work in progress mlḃ api
using local copy of lahman

+++ Baseball1/Dtos/PositionFieldingSummaryDto.cs
@@ -6,8 +6,8 @@
         public string Pos { get; set; } = null!;
         public short? G { get; set; }
         public short? Gs { get; set; }
-        public short? InnOuts { get; set; }
-        public short? Po { get; set; }
+        public int? InnOuts { get; set; }
+        public int? Po { get; set; }
         public short? A { get; set; }
         public short? E { get; set; }
         public short? Dp { get; set; }
